refactor: remove `checkConversationExists` and use `findConversationByInboxIds` directly
